Don't unlink the temp file if it is already gone.
git-svn-id: http://svn.textmate.org/trunk@11743 dfb7d73b-c2ec-0310-8fea-fb051d288c6d

+++ Support/lib/tm/tempfile.rb
@@ -43,13 +43,13 @@ module TextMate
             res = yield(file)
 
             file.close
-            File.unlink(file.path)
+            File.unlink(file.path) if File.exists?(file.path)
             
             return res
           else
             at_exit {
               file.close
-              File.unlink(file.path)              
+              File.unlink(file.path) if File.exists?(file.path)
             }
             return file
           end
